+++ exhibitera/apps/js/exhibitera_app_common.js
@@ -23,7 +23,7 @@ export const config = {
   id: 'TEMP ' + String(new Date().getTime()),
   remoteDisplay: false, // false == we are using the webview app, true == browser
   serverAddress: '',
-  softwareVersion: 5.2,
+  softwareVersion: 5.1,
   standalone: false, // false == we are using Hub
   updateParser: null, // Function used by readUpdate() to parse app-specific updates
   uuid: ''